FIX #19: Increasing hierarchy of Logger: Now it has it's own special place in the project
