+++ modules/infrastructure/src/main/scala/leapfin/infrastructure/stream/utils/search/logger/algebra/Logger.scala
@@ -1,0 +1,9 @@
+package leapfin.infrastructure.stream.utils.search.logger.algebra
+
+import leapfin.infrastructure.stream.utils.search.logger.domain.MatchResultByThread
+
+trait Logger {
+  def feed: MatchResultByThread => Unit
+
+  def printSummarization: Unit
+}
