+++ modules/infrastructure/src/main/scala/leapfin/infrastructure/stream/utils/search/logger/domain/MatchResultByThread.scala
@@ -1,0 +1,5 @@
+package leapfin.infrastructure.stream.utils.search.logger.domain
+
+import leapfin.lemos.word_matcher.algebra.WordMatcher.MatchResult
+
+case class MatchResultByThread(matchResult: MatchResult, threadId: ThreadId)
