+++ modules/infrastructure/src/main/scala/leapfin/infrastructure/stream/utils/search/logger/domain/package.scala
@@ -1,0 +1,5 @@
+package leapfin.infrastructure.stream.utils.search.logger
+
+package object domain {
+  type ThreadId = Long
+}
